refactor - move out habit requests into own module
