+++ src/habits.model.ts
@@ -11,13 +11,6 @@ export interface IHabitMeta {
   type?: string;
 }
 
-export interface IServerHabitData {
-  name: string;
-  _id: string;
-  history: string[];
-  type?: string;
-}
-
 export type IDateHistory = DateStr[];
 
 // TODO: will need to represent date otherwise
